Refactor decorator method, and folderstructure
