+++ backend/src/main/java/com/example/backend/pattern/Decorator/aspect/ImportShipmentServiceLoggingAspect.java
@@ -1,4 +1,4 @@
-package com.example.backend.aspect;
+package com.example.backend.pattern.Decorator.aspect;
 
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.annotation.AfterReturning;
