+++ frontend/src/Hooks/ModalExport/ModalExport.jsx
@@ -19,7 +19,7 @@ import {
   Snackbar,
   Alert
 } from "@mui/material";
-import ExportManagerFacade from "../../Service/ExportManagerFacade"; // Adjust the import path
+import ExportManagerFacade from "../../DesignPatterns/Facade/ExportManagerFacade"; // Adjust the import path
 
 const ModalExport = ({ open, onClose, onSubmit, shipment, setNotification }) => {
   const [orders, setOrders] = useState([]);
