+++ frontend/src/Pages/Report/Report.jsx
@@ -6,7 +6,7 @@ import {
   Modal, Fade, Box, Snackbar, Alert, Select, MenuItem
 } from "@mui/material";
 import TableReport from "../../Hooks/TableReport.jsx/TableReport";
-import ReportManagerFacade from "../../Service/ReportManagerFacade"; // Adjust the import path
+import ReportManagerFacade from "../../DesignPatterns/Facade/ReportManagerFacade"; // Adjust the import path
 
 const style = {
   position: "absolute",
